benchmarks/requirements.txt
google-cloud-storage==1.10.0
pyspark==2.3.1
pyyaml==3.13
mock==2.0.0
